+++ src/Encamina.Enmarcha.AI.OpenAI.Azure/AzureOpenAIOptions.cs
@@ -11,7 +11,7 @@ namespace Encamina.Enmarcha.AI.OpenAI.Azure;
 /// <summary>
 /// Configuration options for Azure OpenAI service connection.
 /// </summary>
-public sealed class AzureOpenAIOptions : OpenAIOptions
+public sealed class AzureOpenAIOptions : OpenAIOptionsBase
 {
     /// <summary>
     /// Gets the Azure OpenAI API service version.
@@ -19,10 +19,28 @@ public sealed class AzureOpenAIOptions : OpenAIOptions
     public OpenAIClientOptions.ServiceVersion ServiceVersion { get; init; } = OpenAIClientOptions.ServiceVersion.V2024_02_15_Preview;
 
     /// <summary>
-    /// Gets the <see cref="System.Uri "/> for an LLM resource (like OpenAI). This should include protocol and host name.
+    /// Gets the <see cref="Uri "/> for an LLM resource (like OpenAI). This should include protocol and host name.
     /// </summary>
     [Required]
     [Uri]
     public Uri Endpoint { get; init; }
-}
 
+    /// <summary>
+    /// Gets the key credential used to authenticate to an LLM resource.
+    /// This property is required if property <see cref="UseTokenCredentialAuthentication"/> is <c>false</c>.
+    /// </summary>
+    [RequiredIf(nameof(UseTokenCredentialAuthentication), false)]
+    [NotEmptyOrWhitespace]
+    public string Key { get; init; }
+
+    /// <summary>
+    /// Gets a value indicating whether Token Credential authentication should be used.
+    /// If set to true, the value of the <see cref="Key"/> property is ignored.
+    /// </summary>
+    public bool UseTokenCredentialAuthentication { get; init; }
+
+    /// <summary>
+    /// Gets the token credentials options to authenticate to an LLM resource.
+    /// </summary>
+    public TokenCredentialsOptions TokenCredentialsOptions { get; init; }
+}